feat: Add delete-app command
Added a new command "delete-app" to the boneless CLI tool. This command allows users to delete an app that was previously created. The command takes the app name as an argument and deletes the corresponding app folder from the internal directory. If the app folder does not exist, an error message is displayed. Upon successful deletion, a confirmation message is printed.